adk/research_upgrade/requirements.txt
google-generativeai==0.3.2
arxiv==2.0.0
python-dotenv==1.0.1
requests==2.31.0
asyncio==3.4.3